requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
jinja2>=3.1.0
aiofiles>=23.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
email-validator>=2.0.0
python3-saml>=1.15.0
starlette-saml>=1.0.0
